Refactor field_decoder into schema package
