+++ README.md
@@ -366,8 +366,8 @@ In short: think of `ParsedMatch` as one container holding both **per-player summ
 |---|---|
 | `binary/reader.py` | `BitReader` — LSB-first bit reading, varint decoding, all binary primitives |
 | `binary/stream.py` | `DemoStream` — outer message loop, Snappy decompression, magic check |
-| `schema/sendtable.py` | Schema layer — serializer + field tree parsed from `CDemoSendTables` |
-| `schema/field_decoder.py` | Type-dispatch decoders including quantized floats |
+| `schema/sendtable/` | Schema layer — serializer + field tree parsed from `CDemoSendTables` |
+| `schema/field_decoder/` | Type-dispatch decoders including quantized floats |
 | `schema/field_path/` | Huffman-coded field path ops for addressing into the serializer tree |
 | `schema/field_state.py` | Nested mutable field-value tree for entity state storage |
 | `schema/field_reader.py` | Field decoder dispatch and entity field reading |
