+++ docs/reference/field_decoder.md
@@ -22,7 +22,7 @@ def find_decoder(field: _FieldLike) -> FieldDecoder
 
 Return the appropriate decoder for the given field.
 
-Source: [src/gem/schema/field_decoder.py:403](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder.py#L403)
+Source: [src/gem/schema/field_decoder/type_resolver.py:124](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder/type_resolver.py#L124)
 
 ## `gem.schema.field_decoder.find_decoder_by_base_type`
 
@@ -34,7 +34,7 @@ def find_decoder_by_base_type(base_type: str) -> FieldDecoder
 
 Return a decoder for a base type string without field context.
 
-Source: [src/gem/schema/field_decoder.py:435](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder.py#L435)
+Source: [src/gem/schema/field_decoder/type_resolver.py:156](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder/type_resolver.py#L156)
 
 ## `gem.schema.field_decoder.QuantizedFloatDecoder`
 
@@ -46,7 +46,7 @@ class QuantizedFloatDecoder
 
 Decoder for Source 2 quantized floats (CNetworkedQuantizedFloat).
 
-Source: [src/gem/schema/field_decoder.py:66](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder.py#L66)
+Source: [src/gem/schema/field_decoder/quantized_float.py:19](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder/quantized_float.py#L19)
 
 #### Methods
 
@@ -56,4 +56,4 @@ Signature: `def QuantizedFloatDecoder.decode(self, r: BitReader) -> float`
 
 Read and decode one quantized float from r.
 
-Source: [src/gem/schema/field_decoder.py:180](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder.py#L180)
+Source: [src/gem/schema/field_decoder/quantized_float.py:131](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_decoder/quantized_float.py#L131)
